File.Format.Dot

Description

Documentation

data DotGraph n :: * -> *

The internal representation of a generalised graph in Dot form.

data GraphID :: *

A polymorphic type that covers all possible ID values allowed by Dot syntax. Note that whilst the ParseDot and PrintDot instances for String will properly take care of the special cases for numbers, they are treated differently here.

Constructors

Str Text

Num Number

Parse a DOT file

dotParse :: Text -> Either String (DotGraph GraphID)

Parses the Text stream from a DOT file.

Get useful representations from the DOT file

dotChildMap :: Ord n => DotGraph n -> Map n (Set n)

Takes a DotGraph parse result and constructs a mapping from a node to it's children.

dotParentMap :: Ord n => DotGraph n -> Map n (Set n)

Takes a DotGraph parse result and constructs a mapping from a node to it's parents.

dotNodeSet :: Ord n => DotGraph n -> Set n

Takes a DotGraph parse result and returns a set of unique node identifiers.

dotEdgeSet :: Ord n => DotGraph n -> Set (n, n)

Takes a DotGraph parse result and returns a set of unique edge identifiers.

toIdentifier :: GraphID -> String

Intelligently render a GraphID to a String for output.