File.Format.Fastc

Description

Functions for for parsing FASTC files into a naive sequence form.

Documentation

type CharacterSequence = Vector (NonEmpty Symbol)

Indexed sequences of Symbols with possible abiguity at an index

type FastcParseResult = NonEmpty FastcSequence

Unconverted result of a fastc parse

data FastcSequence

Pairing of taxa label with an unconverted sequence

Constructors

FastcSequence

Fields

fastcLabel :: Identifier

fastcSymbols :: CharacterSequence

type Identifier = String

Unique identifier for a taxa

type Symbol = String

Component of a phylogenetic character

fastcStreamParser :: (MonadParsec e s m, Token s ~ Char) => m FastcParseResult

Consumes a stream of Chars and parses the stream into a FastcParseResult