File.Format.Newick

Description

The Newick file format was developed by an informal committee meeting at Newick's seafood restaurant. The grammar definition of the Newick format was never formally specified, but Gary Olsen's interpretation of the original newick format has been documented here: http://evolution.genetics.washington.edu/phylip/newick_doc.html

After over two decades of informal usage, the Extended Newick file format was proposed in a BCM publication which allowed node labels to be non- unique and merged to a single node with shared ancestors and descendants. This allowed for easy manual annotating of phylogentic trees.

Another half decade later, the Forest Extended Newick was proposed by Professor Wheeler to model collections of disjoint phylogenetic trees. This new format allowed grouping many Extended Newick trees into a forest to be analyzed collectively.

This parser correctly parses the Newick file format, and the super set Extended Newick filed format, and the new Forest Extended Newick format.

Documentation

type NewickForest = NonEmpty NewickNode

One or more trees in a "Phylogenetic Forest".

data NewickNode

A node in a "Phylogenetic Forest"

isLeaf :: NewickNode -> Bool

Determines whether a given NewickNode is a leaf node in the tree.

newickNode :: [NewickNode] -> Maybe String -> Maybe Double -> Maybe NewickNode

Smart constructor for a NewickNode preseriving the invariant:

null nodes ==> isJust . label

newickStreamParser :: (MonadParsec e s m, Token s ~ Char) => m (NonEmpty NewickForest)

Parses an entire stream into a zero or more NewickForests.

renderNewickForest :: NewickForest -> String

Renders the NewickForest to a String. If the forest contains a DAG with in-degree greater than one, then the shared subtree in a DAG will be rendered multiple times.