File.Format.Nexus

Description

The Nexus file format was developed by an committee consisting of members who had no concept of language design. Consequently the Nexus file format is needlessly complex.

This parser works barely on a subset of the Nexus specification's ambiguously defined features.

Documentation

type AlphabetSymbol = String

Represents a symbol in the character alphabet.

type AmbiguityGroup = [AlphabetSymbol]

Represents an ambiguity group of symbols in the character alphabet. Should always be a non empty list.

type Character = Maybe (Vector AmbiguityGroup)

Character is Maybe, because some chars might not be present for some taxa if unaligned, multiple characters are treated as a single "character", so V.length >= 1 if aligned, each character is a "character", so V.length == 1

data CharacterMetadata

Metadata record type for a given character.

Constructors

CharacterMetadata

Fields

name :: String

isAligned :: Bool

charType :: CharDataType

alphabet :: AmbiguityGroup

ignored :: Bool

costM :: Maybe TCM

additive :: Bool

weight :: Int

data CharDataType

The types of data which can be present in a Nexus file. This type might get scrapped and another type imported from different module, or preserved but moved to another module.

Constructors

Standard

DNA

RNA

Nucleotide

Protein

Continuous

data Nexus

The collection of information extracted from blocks in the Nexus file.

Constructors

Nexus

Fields

sequences :: Sequences

nexusForests :: [NewickForest]

type Sequence = Vector Character

This is a Vector, because Character may have length /= 1 (see explanation at Character)

type Sequences = (TaxonSequenceMap, Vector CharacterMetadata)

Mapping of taxa names to character data and a vector corresponding to each character's metadata.

type TaxonIdentifier = String

Represents the unique identifier of a taxon.

type TaxonSequenceMap = Map TaxonIdentifier Sequence

Mapping of taxon identifiers (typically names) to thier corresponding sequence data.

nexusStreamParser :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => m Nexus

Parses the entirety of a stream consisting of a single Nexus file resulting a Nexus.