File.Format.TNT

Description

Parser for the Henning or TNT file format. Parses a useful subset of the file specification.

Documentation

data CharacterMetaData

The metadata of a character specifying the attributes of the character specified in the file.

Default 'CharacterMetaData values:

>>> defaultMetaData
CharMeta
  { characterName   = ""
  , characterStates = mempty
  , additive        = False
  , active          = True
  , sankoff         = False
  , weight          = 1
  , steps           = 1
  , costTCM         = Nothing
  }

Constructors

CharMeta

Fields

characterName :: String

characterStates :: Vector String

additive :: Bool

active :: Bool

sankoff :: Bool

weight :: Int

steps :: Int

costTCM :: Maybe (Matrix Double)

data LeafyTree a

A rose tree which only contains data at the leaf nodes.

Constructors

Leaf a

Branch [LeafyTree a]

type TaxonInfo = (TaxonName, TaxonSequence)

The sequence information for a taxon within the TNT file's XREAD command. Contains the TaxonName and the naive TaxonSequence

data TntCharacter

Different character types are deserialized from sequences segments. After all segments are collected they are de-interleaved into a single TaxonSequence.

Constructors

Continuous TntContinuousCharacter

Discrete TntDiscreteCharacter

Dna TntDnaCharacter

Protein TntProteinCharacter

type TntResult = Either TreeOnly WithTaxa

The result of parsing a TNT file. A file can contain either only tree data with labeled leaf nodes or a collection of taxa sequences with coresponsing metadata and possibly corresponding tree data.

type TRead = NonEmpty TReadTree

Parse result of an internal TREAD command. Specifies a nonempty forest of trees, where the leaf node set must be validated againtst the taxa set from a XREAD command.

type TReadTree = LeafyTree NodeType

A tree with data only at the leaf nodes. Leaf nodes contain different criteria for matching the node against a given taxa from the taxa set.

type TreeOnly = TRead

The possible parse result when the file contains only tree data.

data NodeType

Multiple possible criteria for matching leaf nodes to taxa from the taxa set.

Constructors

Index Int

Name String

Prefix String

data WithTaxa

The possible parse result when the file contains taxa sequences. trees represents a (possibly empty) forest where each tree in the forest must have the complete taxa set as it's leaf node set.

Constructors

WithTaxa

Fields

sequences :: Vector TaxonInfo

charMetaData :: Vector CharacterMetaData

trees :: [LeafyTree TaxonInfo]

type TaxonName = String

The name of a taxon in a TNT file's XREAD command.

type TaxonSequence = [TntCharacter]

The naive sequence of a taxon in a TNT files' XREAD command.

type TntContinuousCharacter = Maybe Double

A TntContinuousCharacter is an real valued character. Continuous characters can have negative values. Continuous values are serialized textually as decimal values or integral values, scientific notation supported.

data TntDiscreteCharacter

A TntDiscreteCharacter is an integral value in the range '[0..62]'. Discrete values are serialized textualy as one of the 64 values: '[0..9] ++ ['A'..'B'] ++ ['a'..'z'] ++ "-?"'. Missing '?' represents the empty ambiguity group. Each value coresponds to it's respective bit in the Int64. Ambiguity groups are represented by Int64 values with multiple set bits.

data TntDnaCharacter

A TntDnaCharacter represents a nucleotide ('ACGT') as an integral value in the range '[0..3]' respectively. If gap characters ('-') are treated as a fifth state then values the additional value '4' is added to the integral range. Discrete values are serialized textualy as the DNA IUPAC codes case-insensitively, along with '-' & '?' characters representing gap or missing data respecitively. Gap represents an ambiguity group of all possible proteins unless gaps are treated as a fifth state. Missing represents the empty ambiguity group.

data TntProteinCharacter

A TntProteinCharacter is an integral value in the range '[0..20]'. Discrete values are serialized textualy as the protein IUPAC codes case-insensitively, along with '-' & '?' characters representing gap or missing data respecitively. Missing represents the empty ambiguity group.

tntStreamParser :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => m TntResult

Parses the contents of a TNT file stream into a TntResult. A file stream can contain either:

Only a forest of trees with labeled leaf nodes and the set of leaf nodes is the same across the forest.

A collection of taxa sequences with coresponsing metadata and possibly corresponding forest of trees whose leaf sets are equal to the taxa set.