File.Format.TransitionCostMatrix

Description

Functions for for parsing TCM files into an alphabet and native matrix form.

Documentation

data TCM

The results of a TCM file consisting of

A custom alphabet of Symbols

A matrix consisting of the transition costs between symbols

The following equality will hold for an TCM:

(length . customAlphabet) tcm == (nrows . transitionCosts) tcm && (length . customAlphabet) tcm == (ncols . transitionCosts) tcm

Note that the transitionCosts does not need to be a symetic matrix nor have identity values on the matrix diagonal.

Constructors

TCM

Fields

customAlphabet :: NonEmpty String

The custom alphabet of Symbols for which the TCM matrix is defined

transitionCosts :: Matrix Double

The cost to transition between any two symbols, square but not necessarily symetric

tcmStreamParser :: (MonadParsec e s m, Token s ~ Char) => m TCM

Parses the entirety of a stream producing a TCM result. The result will contain an Alphabet with no duplicate elements and a square Matrix with dimension (n+1) x (n+1) where n is the length of the Alphabet.