File.Format.VertexEdgeRoot

Description

Functions for for parsing VER files into a collection of the vertex set, edge set, and root set representing a "Phylogenetic Forest".

Documentation

data EdgeInfo

Connection between two nodes in the graph along with the distance of the connection. Edges are interpred as bidirectional when read in and given direction when interpreted relative to the root of the tree.

type EdgeLength = Maybe Double

The possibly calculated distance between two nodes in the graph

data VertexEdgeRoot

Collection of Vericies, roots, and edges representing a "Phylogenetic Forest"

Constructors

VER

Fields

vertices :: Set VertexLabel

edges :: Set EdgeInfo

roots :: Set VertexLabel

type VertexLabel = String

A textual identifier for a node in the graph

connectedVertex :: VertexLabel -> EdgeInfo -> Maybe VertexLabel

For a given vertex, attempts to get the connected vertex from the EdgeInfo. If the input vertex was present in the EdgeInfo, returns 'Just v' where v is the corresponsing VertexLabel. If the input vertex was not present in the EdgeInfo.

edgeOrigin :: EdgeInfo -> VertexLabel

Extract the origin of the directed edge

edgeTarget :: EdgeInfo -> VertexLabel

Extract the destination of the directed edge

edgeLength :: EdgeInfo -> EdgeLength

Extract the possibly present edge length

verStreamParser :: (FoldCase (Tokens s), MonadParsec e s m, Token s ~ Char) => m VertexEdgeRoot

Reads two vertex sets and an edge set, conditionally infers the root set when vertex sets are unlabeled. Ensures that the elements of the root set are not connected in the forest. Ensures that the rooted trees in the forest do not contain cycles.